Backend/requirements.txt
argcomplete==3.1.4
asgiref==3.7.2
astroid==3.0.2
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
cliapp==1.20180812.1
click==8.1.6
cloud-init==24.4.1
cmdtest==0.32+git
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
dill==0.3.8
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
django-anymail==12.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
duplicity==2.1.4
fasteners==0.18
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
inflection==0.5.1
isort==5.6.4
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
logilab-common==1.9.8
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mypy-extensions==1.0.0
netaddr==0.8.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
psycopg==3.2.6
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pylint==3.0.3
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
qrcode==8.0
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==11.6.0
systemd-python==235
toml==0.10.2
tomlkit==0.12.4
ttystatus==0.38
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
wheel==0.42.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zombie-imp==0.0.2
